Dev01

+++ app/models/attendance.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Date, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, Date, DateTime, ForeignKey, BigInteger
 from datetime import datetime
 from app.db.database import Base
 
@@ -6,7 +6,7 @@ class Attendance(Base):
     __tablename__ = "attendance"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"))
+    user_id = Column(BigInteger, ForeignKey("users.id"))
 
     date = Column(Date)
     check_in = Column(DateTime, default=datetime.utcnow)
